refactor: Apply team code conventions and formatting rules
