+++ onboard-app-guilherme/.gitignore
@@ -35,3 +35,7 @@ yarn-error.*
 
 # typescript
 *.tsbuildinfo
+
+# native
+android/
+ios/
