+++ onboard-app-guilherme/App.tsx
@@ -1,23 +1,14 @@
 // src/App.tsx
 
-import { StatusBar } from 'expo-status-bar';
-import { StyleSheet, View } from 'react-native';
-import { HelloWorld } from './src/components/HelloWorld/hello-world';
+import { StatusBar } from "expo-status-bar";
+import { View } from "react-native";
+import { HelloWorld } from "./src/components/hello-world";
 
 export default function App() {
-  return (
-    <View style={styles.container}>
-      <HelloWorld />
-      <StatusBar style="auto" />
-    </View>
-  );
+	return (
+		<View>
+			<HelloWorld />
+			<StatusBar style="auto" />
+		</View>
+	);
 }
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: '#fff',
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-});
