+++ onboard-app-guilherme/package.json
@@ -18,6 +18,7 @@
   },
   "devDependencies": {
     "@babel/core": "^7.25.2",
+    "@biomejs/biome": "^2.2.2",
     "@types/react": "~19.0.10",
     "@types/react-native": "^0.73.0",
     "@types/styled-components-react-native": "^5.2.5",
