--- onboard-app-guilherme/src/components/HelloWorld/hello-world.tsx
@@ -1,5 +1,0 @@
-import { HelloWorldText } from './styles';
-
-export function HelloWorld() {
-  return <HelloWorldText>Hello World</HelloWorldText>;
-}
