--- onboard-app-guilherme/src/components/HelloWorld/styles.ts
@@ -1,7 +1,0 @@
-import styled from 'styled-components/native';
-
-export const HelloWorldText = styled.Text`
-  font-size: 28px;
-  font-weight: bold;
-  color: #333;
-`;
