+++ onboard-app-guilherme/src/components/hello-world/index.tsx
@@ -1,0 +1,5 @@
+import { HelloWorldText } from "./styles";
+
+export function HelloWorld() {
+	return <HelloWorldText>Hello World</HelloWorldText>;
+}
