+++ onboard-app-guilherme/src/components/hello-world/styles.ts
@@ -1,0 +1,9 @@
+import styled from "styled-components/native";
+
+export const HelloWorldText = styled.Text`
+  font-size: 28px;
+  font-weight: bold;  
+  color: #333;
+  text-align: center;
+  
+`;
